Finished the mapping algorithm to use a BDF algo & I've ensured it iterates over each line at every station.
I'm using a priority queue to work through the best performing nodes at first. I also set a max limit of the research as the complex structure of the network means it can take hours to find a solution that is unlikely to be better.

I have also created a separate file that fetches from the api and visually displays this content.